Removed unnecessary check for bound method & moevd decorator definitions to utils file
